on the booking note, the date is now todays date, instead of the order date
