+++ src/components/orders/add-order.jsx
@@ -5,23 +5,7 @@ var Units = require("./units.jsx");
 var Warning = require("../close-warning.jsx");
 var DataList = require("./data-list.jsx");
 
-var currentDate = function () {
-    var today = new Date();
-
-    var dd = today.getDate();
-    var mm = today.getMonth()+1;
-    var yyyy = today.getFullYear();
-
-    if (dd < 10){
-        dd='0'+dd
-    }
-    if (mm < 10) {
-        mm='0'+mm
-    }
-
-    var todayDate = yyyy + "-" + mm + "-" + dd;
-    return todayDate;
-}	
+var currentDate = require("../../lib/current-date.js");
 
 var addOrder = React.createClass({
 	getInitialState: function() {
